py/requirements.txt
pykakasi
python_dotenv
uvicorn
fastapi
